Add 02a-nsxt-tf-var stage and additional FW options
Add a new 02a-nsxt-tf-var stage, this can be implemented instead of the
original 02a-nsxt.

02a-nsxt-tf-var is similar to 02a-nsxt but uses extended options for all
the resources in the stage. This is in preparation for other stages that
should expose all options available in the NSX-T resources.

To implement per app resources, for example groups or distributed
firewall policies, there is a single Terraform variable, that accepts a
list of resource configurations.

Additionally, add more options available to the
nsxt-distributed-firewall-policy module, so we can expose them in the
02a-nsxt-tf-var module.